+++ code/handler.c
@@ -385,7 +385,7 @@ void affect_total(struct char_data* ch) {
     if (ch->equipment[i]) {
       for (j = 0; j < MAX_OBJ_AFFECT; j++) {
         affect_modify(ch, ch->equipment[i]->affected[j].location,
-          (long)ch->equipment[i]->affected[j].modifier,
+          ch->equipment[i]->affected[j].modifier,
           ch->equipment[i]->obj_flags.bitvector, 0);
       }
     }
@@ -401,7 +401,7 @@ void affect_total(struct char_data* ch) {
     if (ch->equipment[i]) {
       for (j = 0; j < MAX_OBJ_AFFECT; j++) {
         affect_modify(ch, ch->equipment[i]->affected[j].location,
-          (long)ch->equipment[i]->affected[j].modifier,
+          ch->equipment[i]->affected[j].modifier,
           ch->equipment[i]->obj_flags.bitvector, 1);
       }
     }
@@ -816,8 +816,8 @@ void equip_char(struct char_data* ch, struct obj_data* obj, int pos) {
   }
 
   for (j = 0; j < MAX_OBJ_AFFECT; j++) {
-    affect_modify(ch, obj->affected[j].location,
-      (long)obj->affected[j].modifier, obj->obj_flags.bitvector, 1);
+    affect_modify(ch, obj->affected[j].location, obj->affected[j].modifier,
+      obj->obj_flags.bitvector, 1);
   }
 
   if (GET_ITEM_TYPE(obj) == ITEM_WEAPON) {
